python_worker/pyproject.toml
[tool.poetry]
name = "python-worker"
version = "0.1.0"
description = ""
authors = ["plozovik <plozovikov@gmail.com>"]
readme = "README.md"
packages = [{ include = "python_worker" }]

[tool.poetry.dependencies]
python = "^3.9"
temporalio = "^1.1.0"
opentelemetry-exporter-otlp-proto-grpc = "1.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
